refactor: use DashboardModel type of schema
