+++ apps/web/src/lib/helper/config-data-helper.ts
@@ -1,6 +1,7 @@
 import { cloneDeep, find } from 'lodash';
 
 import type { CostQuerySetModel } from '@/schema/cost-analysis/cost-query-set/model';
+import type { DashboardModel } from '@/schema/dashboard/_types/dashboard-type';
 import type { WorkspaceModel } from '@/schema/identity/workspace/model';
 import type { MetricExampleModel } from '@/schema/inventory/metric-example/model';
 import { i18n } from '@/translations';
@@ -20,7 +21,6 @@ import type { FavoriteConfig, FavoriteItem } from '@/common/modules/favorites/fa
 import type { RecentConfig } from '@/common/modules/navigations/type';
 
 import { COST_EXPLORER_ROUTE } from '@/services/cost-explorer/routes/route-constant';
-import type { DashboardModel } from '@/services/dashboards/types/dashboard-api-schema-type';
 
 type Config = FavoriteConfig & RecentConfig;
